requirements.txt
# requirements.txt
langchain
langchain-google-genai
streamlit
pandas
numpy
plotly
sqlalchemy
python-dotenv
openpyxl
chromadb
